remove getUserPermissionsById() use case, remove FilePermissions from File.ts, rename FileUserPermissions.ts to FilePermissions.ts
